[C2RKDump]: Fix debug frame timing crash
error log:
ubsan: mul-overflow by 0xe49ca0ef

Type: Fix
Redmine ID: N/A
Associated modifications: N/A
Test: N/A

+++ component/osal/C2RKDump.cpp
@@ -57,7 +57,7 @@ const char *toStr_RawType(uint32_t type) {
 int64_t getCurrentTimeMillis() {
     struct timeval tv;
     gettimeofday(&tv, nullptr);
-    return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+    return (int64_t)tv.tv_sec * 1000 + tv.tv_usec / 1000;
 }
 
 C2RKDump::C2RKDump()
